make Pagination have uncomposed version
