+++ packages/react-storage/src/components/StorageBrowser/subcomponents/PaginationControl/CurrentPage.tsx
@@ -19,7 +19,7 @@ export const CurrentPage = <T extends ButtonElementProps>({
       disabled={isDisabled}
       aria-current="page"
     >
-      {children}
+      {children ?? '1'}
     </Button>
   );
 };
