+++ packages/react-storage/src/components/StorageBrowser/subcomponents/PaginationControl/PaginationControl.tsx
@@ -1,30 +1,37 @@
 import * as React from 'react';
 import { NavElementProps } from '@aws-amplify/ui-react/internal';
-import { useElement } from '../../context/elements';
 import { NextButton } from './NextButton';
 import { PreviousButton } from './PreviousButton';
+import { Container } from './Container';
 import { CurrentPage } from './CurrentPage';
 import { Item } from './Item';
 import { List } from './List';
 
 const Pagination = <T extends NavElementProps>({
-  ariaLabel = 'Pagination',
-  className: _className,
+  ariaLabel,
+  className,
   children,
   ...rest
 }: T): JSX.Element => {
-  const Nav = useElement('Nav');
-  const baseClassName = 'amplify-storagebrowser__pagination';
-  const className = _className ?? baseClassName;
-
   return (
-    <Nav {...rest} aria-label={ariaLabel} className={className}>
-      {children}
-    </Nav>
+    <Container ariaLabel={ariaLabel} className={className} {...rest}>
+      <List>
+        <Item>
+          <PreviousButton />
+        </Item>
+        <Item>
+          <CurrentPage />
+        </Item>
+        <Item>
+          <NextButton />
+        </Item>
+      </List>
+    </Container>
   );
 };
 
 const PaginationControl = Object.assign(Pagination, {
+  Container,
   Item,
   List,
   NextButton,
